requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-Login==0.6.2
Flask-WTF==1.1.1
python-dotenv==1.0.0
Werkzeug==2.3.6
SQLAlchemy==2.0.19
PyMySQL==1.1.0
cryptography==41.0.3
dnspython==2.6.1
email_validator==2.2.0
idna==3.7